update container methodsand constants
