+++ src/Constants/Index.js
@@ -1,10 +1,10 @@
-export const getMeals = state => state.products;
-export const getMealsPending = state => state.pending;
-export const getMealsError = state => state.error;
+export const getProducts = state => state.products;
+export const getProductsPending = state => state.pending;
+export const getProductsError = state => state.error;
 export const getImg = state => state.strMealThumb;
 export const getArea = state => state.strArea;
 export const getName = state => state.strMeal;
-export const getInstructions = state => state.strInstructions;
+export const getIntructions = state => state.strInstructions;
 export const getCategoriesList = state => state.categories;
 export const categoryName = category => category.strCategory;
 export const categoryDetails = category => category.strCategoryDescription;
@@ -12,7 +12,8 @@ export const categoryImage = category => category.strCategoryThumb;
 
 export const getIngredient = meal => {
   const result = [];
-  for (let i = 1; i <= 20; i += 1) {
+  // eslint-disable-next-line no-plusplus
+  for (let i = 1; i <= 30; i++) {
     if (meal[`strIngredient${i}`] === '') {
       break;
     }
